requirements.txt
colorama        # styling
inquirerpy      # option selection
pyfiglet        # logo
argon2-cffi     # login encryption
cryptography    # CSV encryption 
pytest          # testing
keyring